+++ wdsp/ammod.hpp
@@ -40,21 +40,21 @@ public:
     int run;
     int mode;
     int size;
-    double* in;
-    double* out;
-    double c_level;
-    double a_level;
-    double mult;
+    float* in;
+    float* out;
+    float c_level;
+    float a_level;
+    float mult;
 
-    static AMMOD* create_ammod(int run, int mode, int size, double* in, double* out, double c_level);
+    static AMMOD* create_ammod(int run, int mode, int size, float* in, float* out, float c_level);
     static void destroy_ammod (AMMOD *a);
     static void flush_ammod (AMMOD *a);
     static void xammod (AMMOD *a);
-    static void setBuffers_ammod (AMMOD *a, double* in, double* out);
+    static void setBuffers_ammod (AMMOD *a, float* in, float* out);
     static void setSamplerate_ammod (AMMOD *a, int rate);
     static void setSize_ammod (AMMOD *a, int size);
     // TXA Properties
-    static void SetAMCarrierLevel (TXA& txa, double c_level);
+    static void SetAMCarrierLevel (TXA& txa, float c_level);
 };
 
 } // namespace WDSP
