+++ wdsp/bandpass.hpp
@@ -57,27 +57,27 @@ public:
     int size;
     int nc;
     int mp;
-    double* in;
-    double* out;
-    double f_low;
-    double f_high;
-    double samplerate;
+    float* in;
+    float* out;
+    float f_low;
+    float f_high;
+    float samplerate;
     int wintype;
-    double gain;
+    float gain;
     FIRCORE *p;
 
-    static BANDPASS *create_bandpass (int run, int position, int size, int nc, int mp, double* in, double* out,
-        double f_low, double f_high, int samplerate, int wintype, double gain);
+    static BANDPASS *create_bandpass (int run, int position, int size, int nc, int mp, float* in, float* out,
+        float f_low, float f_high, int samplerate, int wintype, float gain);
     static void destroy_bandpass (BANDPASS *a);
     static void flush_bandpass (BANDPASS *a);
     static void xbandpass (BANDPASS *a, int pos);
-    static void setBuffers_bandpass (BANDPASS *a, double* in, double* out);
+    static void setBuffers_bandpass (BANDPASS *a, float* in, float* out);
     static void setSamplerate_bandpass (BANDPASS *a, int rate);
     static void setSize_bandpass (BANDPASS *a, int size);
-    static void setGain_bandpass (BANDPASS *a, double gain, int update);
-    static void CalcBandpassFilter (BANDPASS *a, double f_low, double f_high, double gain);
+    static void setGain_bandpass (BANDPASS *a, float gain, int update);
+    static void CalcBandpassFilter (BANDPASS *a, float f_low, float f_high, float gain);
     // RXA Prototypes
-    static void SetBandpassFreqs (RXA& rxa, double f_low, double f_high);
+    static void SetBandpassFreqs (RXA& rxa, float f_low, float f_high);
     static void SetBandpassNC (RXA& rxa, int nc);
     static void SetBandpassMP (RXA& rxa, int mp);
     // TXA Prototypes
